library: Add x and y axis independent sprite scaling
I added this feature for a game I'm personally working on and am opening
this if it is a desired upstream feature.

Feature allows for X and Y axis sprite scaling, presently sprite scaling
affects both axis. This additionally will allow for horizontal and
vertical flip of sprites by nature of independent axis scaling.